# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.ConsumptionEstimate:
    properties:
      area_ha:
        type: number
      fuel_resources:
        items:
          $ref: '#/definitions/domain.FuelStock'
        type: array
      items:
        items:
          $ref: '#/definitions/domain.ConsumptionEstimateItem'
        type: array
    type: object
  domain.ConsumptionEstimateItem:
    properties:
      category:
        type: string
      quantity:
        type: number
      quantity_per_unit:
        type: number
      resource_id:
        type: integer
      resource_name:
        type: string
      unit:
        type: string
    type: object
...
  domain.FieldOperationResourceUsage:
    properties:
      quantity:
        type: number
      resource_id:
        type: integer
    type: object
  domain.FuelStock:
    properties:
      quantity:
        type: number
      resource_id:
        type: integer
      resource_name:
        type: string
      unit:
        type: string
    type: object
...
  domain.OperationType:
    enum:
    - soil_preparation
    - seeding
    - fertilization
    - spraying
    - harvesting
    - irrigation
    type: string
    x-enum-varnames:
    - OperationTypeSoilPreparation
    - OperationTypeSeeding
    - OperationTypeFertilization
    - OperationTypeSpraying
    - OperationTypeHarvesting
    - OperationTypeIrrigation
...
  domain.ReportFieldRow:
    properties:
      area_ha:
        type: number
      cadastral_number:
        type: string
      completed_count:
        type: integer
      estimated_cost:
        type: number
      fuel_used_l:
        type: number
      geometry:
        type: object
      id:
        type: string
      last_operation_at:
        type: string
      name:
        type: string
      operations_count:
        type: integer
      planned_area_ha:
        type: number
      real_cost:
        type: number
      realized_area_ha:
        type: number
    type: object
  domain.ReportFields:
    properties:
      fields_with_operations:
        type: integer
      items:
        items:
          $ref: '#/definitions/domain.ReportFieldRow'
        type: array
      period:
        $ref: '#/definitions/domain.ReportPeriod'
      total_area_ha:
        type: number
      total_fields:
        type: integer
      worked_area_ha:
        type: number
    type: object
...
  domain.ReportNamedCount:
    properties:
      count:
        type: integer
      key:
        type: string
    type: object
...
  domain.ReportOperationRow:
    properties:
      actual_duration_minutes:
        type: integer
      actual_end_at:
        type: string
      actual_start_at:
        type: string
      area_completed_ha:
        type: number
      area_planned_ha:
        type: number
      delay_minutes:
        type: integer
      estimated_cost:
        type: number
      field_id:
        type: string
      field_name:
        type: string
      fuel_used_l:
        type: number
      id:
        type: integer
      implement_name:
        type: string
      machine_hours:
        type: number
      machine_name:
        type: string
      operation_type_name:
        type: string
      operator_name:
        type: string
      planned_end_at:
        type: string
      planned_start_at:
        type: string
      real_cost:
        type: number
      status:
        type: string
      template_name:
        type: string
    type: object
  domain.ReportOperationTypeStat:
    properties:
      area_ha:
        type: number
      completed:
        type: integer
      estimated_cost:
        type: number
      operation_type:
        $ref: '#/definitions/domain.OperationType'
      operation_type_name:
        type: string
      total:
        type: integer
    type: object
  domain.ReportOperations:
    properties:
      by_status:
        items:
          $ref: '#/definitions/domain.ReportNamedCount'
        type: array
      by_type:
        items:
          $ref: '#/definitions/domain.ReportOperationTypeStat'
        type: array
      items:
        items:
          $ref: '#/definitions/domain.ReportOperationRow'
        type: array
      metrics:
        $ref: '#/definitions/domain.ReportOperationsMetrics'
      period:
        $ref: '#/definitions/domain.ReportPeriod'
      timeline:
        items:
          $ref: '#/definitions/domain.ReportTimeBucket'
        type: array
    type: object
  domain.ReportOperationsMetrics:
    properties:
      actual_duration_minutes:
        type: integer
      completed_area_ha:
        type: number
      completed_with_actuals:
        description: 'Valori reale (faza 2): disponibile doar pentru operațiunile
          finalizate cu date reale.'
        type: integer
      estimated_cost:
        type: number
      fields_worked:
        type: integer
      fuel_used_l:
        type: number
      machine_hours:
        type: number
      machines_used:
        type: integer
      on_time_completed:
        type: integer
      operations_canceled:
        type: integer
      operations_completed:
        type: integer
      operations_in_progress:
        type: integer
      operations_planned:
        type: integer
      operations_total:
        type: integer
      operators_used:
        type: integer
      overdue_operations:
        type: integer
      planned_area_ha:
        type: number
      real_cost:
        type: number
      realized_area_ha:
        type: number
    type: object
...
  domain.ReportPeriod:
    properties:
      from:
        type: string
      granularity:
        type: string
      previous_from:
        type: string
      previous_to:
        type: string
      to:
        type: string
    type: object
...
  domain.ReportTimeBucket:
    properties:
      area_ha:
        type: number
      bucket:
        type: string
      canceled:
        type: integer
      completed:
        type: integer
      in_progress:
        type: integer
      planned:
        type: integer
    type: object
...
  domain.StockMovement:
    properties:
      actor_id:
        type: integer
      actor_name:
        type: string
      category:
        type: string
      created_at:
        type: string
      field_crop_id:
        type: integer
      field_operation_id:
        type: integer
      field_operation_label:
        type: string
      id:
        type: integer
      movement_type:
        $ref: '#/definitions/domain.StockMovementType'
      notes:
        type: string
      quantity_delta:
        type: number
      resource_id:
        type: integer
      resource_name:
        type: string
      resulting_quantity:
        type: number
      total_cost:
        type: number
      unit:
        type: string
      unit_cost:
        type: number
    type: object
  domain.StockMovementType:
    enum:
    - in
    - out
    - adjustment
    type: string
    x-enum-varnames:
    - StockMovementIn
    - StockMovementOut
    - StockMovementAdjustment
...
  handlers.completeFieldOperationRequest:
    properties:
      actual_end_at:
        type: string
      area_completed_ha:
        type: number
      consume_from_template:
        type: boolean
      fuel_resource_id:
        type: integer
      fuel_used_l:
        type: number
      machine_hours:
        type: number
      notes:
        type: string
      resources:
        items:
          $ref: '#/definitions/domain.FieldOperationResourceUsage'
        type: array
    type: object
  handlers.completeFieldOperationResponse:
    properties:
      movements:
        items:
          $ref: '#/definitions/domain.StockMovement'
        type: array
      operation: {}
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST pentru managementul mașinilor agricole, operatorilor și asignărilor.
  title: Agri API
  version: "1.0"
paths:
...
  /field-operations/{id}/complete:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID operațiune
        in: path
        name: id
        required: true
        type: integer
      - description: Datele reale ale finalizării
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.completeFieldOperationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.completeFieldOperationResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Finalizare operațiune pe teren
      tags:
      - field-operations
  /field-operations/{id}/consumption-estimate:
    get:
      parameters:
      - description: ID operațiune
        in: path
        name: id
        required: true
        type: integer
      - description: Suprafața realizată (implicit cea planificată)
        in: query
        name: area_ha
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ConsumptionEstimate'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Estimare consum la finalizare
      tags:
      - field-operations
...
  /reports/fields:
    get:
      parameters:
      - description: Data de început (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Data de sfârșit (YYYY-MM-DD)
        in: query
        name: to
        type: string
      - description: Filtru teren
        in: query
        name: field_id
        type: string
      - description: Filtru tip operațiune (soil_preparation, seeding, fertilization,
          spraying, harvesting, irrigation)
        in: query
        name: operation_type
        type: string
      - description: Filtru mașină
        in: query
        name: machine_id
        type: integer
      - description: Filtru operator
        in: query
        name: operator_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ReportFields'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Raport terenuri
      tags:
      - reports
...
  /reports/operations:
    get:
      parameters:
      - description: Data de început (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: Data de sfârșit (YYYY-MM-DD)
        in: query
        name: to
        type: string
      - description: Filtru teren
        in: query
        name: field_id
        type: string
      - description: Filtru tip operațiune (soil_preparation, seeding, fertilization,
          spraying, harvesting, irrigation)
        in: query
        name: operation_type
        type: string
      - description: Filtru mașină
        in: query
        name: machine_id
        type: integer
      - description: Filtru operator
        in: query
        name: operator_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ReportOperations'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Raport operațiuni
      tags:
      - reports
...
securityDefinitions:
  BearerAuth:
    description: 'Introdu token-ul cu prefixul "Bearer ": **Bearer &lt;token&gt;**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"